requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
PyYAML>=6.0
scapy>=2.4.0
jupyter>=1.0.0